cli/new: replace additionalActions with fixed installation per role
