+++ packages/cli/templates/default-backend-module.yaml
@@ -1,8 +1,5 @@
 description: A new backend module that extends an existing backend plugin with additional features
 template: ./default-backend-module
 role: backend-plugin-module
-additionalActions:
-  - install-backend
-  - add-backend
 templateValues:
   moduleVar: '{{ camelCase pluginId }}Module{{ upperFirst ( camelCase moduleId ) }}'
