+++ packages/cli/templates/scaffolder-module.yaml
@@ -5,6 +5,3 @@ params:
   pluginId: scaffolder
 templateValues:
   moduleVar: '{{ camelCase pluginId }}Module{{ upperFirst ( camelCase moduleId ) }}'
-additionalActions:
-  - install-backend
-  - add-backend
